implemented Triangle Rewards.py
This file is used for the reading of Data from Database and for the construction of the list of Themes (5 in total) for each individual client, from their previous purchases. It also has a function where, when the individual Themes of the client are not found, it finds an appropriate Mode Theme of the grand total of all clients.